Generator: move files (nether)
