+++ src/pocketmine/Server.php
@@ -103,7 +103,7 @@ use pocketmine\level\generator\biome\Biome;
 use pocketmine\level\generator\Flat;
 use pocketmine\level\generator\Void;
 use pocketmine\level\generator\Generator;
-use pocketmine\level\generator\hell\Nether;
+use pocketmine\level\generator\nether\Nether;
 use pocketmine\level\generator\normal\Normal;
 use pocketmine\level\generator\normal\Normal2;
 use pocketmine\level\Level;
